Fix activated preregistered patient login bug
There was a bug that would cause a database exception whenver a
preregistered patient tried to login after an HCP approved them. This
was a result of the patient's MID not being set in the log transaction
when a preregistered patient is created. This commit introduces a fix
for this bug.