UC 14: Change trends tab start date
The trends tab in diagnosis trends should use a start date that is 8
weeks prior to the end date. This commit introduces a change to remove
the start date field and parameter and compute the start date based on
the provided end date. It also contains the junit tests associated with
the change.